+++ src/main/java/com/kalanso/event/Repository/AdminRepo.java
@@ -3,5 +3,5 @@ package com.kalanso.event.Repository;
 import com.kalanso.event.Model.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminRepo extends JpaRepository<Admin, Long> {
+public interface AdminRepo extends JpaRepository<Admin, Integer> {
 }
