+++ src/main/java/com/kalanso/event/Repository/ClientRepo.java
@@ -3,5 +3,5 @@ package com.kalanso.event.Repository;
 import com.kalanso.event.Model.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientRepo extends JpaRepository<Client, Long> {
+public interface ClientRepo extends JpaRepository<Client, Integer> {
 }
