+++ src/main/java/com/kalanso/event/Repository/GestionnaireRepo.java
@@ -3,5 +3,5 @@ package com.kalanso.event.Repository;
 import com.kalanso.event.Model.Gestionnaire;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GestionnaireRepo extends JpaRepository<Gestionnaire, Long> {
+public interface GestionnaireRepo extends JpaRepository<Gestionnaire, Integer> {
 }
