+++ src/main/java/com/kalanso/event/Repository/OrgaRepo.java
@@ -3,5 +3,5 @@ package com.kalanso.event.Repository;
 import com.kalanso.event.Model.Organisateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrgaRepo extends JpaRepository<Organisateur, Long> {
+public interface OrgaRepo extends JpaRepository<Organisateur, Integer> {
 }
